+++ src/Schema/AbstractSchema.php
@@ -12,6 +12,14 @@ use Yiisoft\Db\Connection\ConnectionInterface;
 use Yiisoft\Db\Constraint\Constraint;
 use Yiisoft\Db\Constraint\IndexConstraint;
 use Yiisoft\Db\Exception\NotSupportedException;
+use Yiisoft\Db\Schema\Column\BinaryColumnSchema;
+use Yiisoft\Db\Schema\Column\BooleanColumnSchema;
+use Yiisoft\Db\Schema\Column\ColumnSchemaInterface;
+use Yiisoft\Db\Schema\Column\DoubleColumnSchema;
+use Yiisoft\Db\Schema\Column\IntegerColumnSchema;
+use Yiisoft\Db\Schema\Column\JsonColumnSchema;
+use Yiisoft\Db\Schema\Column\StringColumnSchema;
+use Yiisoft\Db\Schema\Column\BigIntColumnSchema;
 
 use function gettype;
 use function is_array;
@@ -365,24 +373,46 @@ abstract class AbstractSchema implements SchemaInterface
     }
 
     /**
-     * Extracts the PHP type from an abstract DB type.
+     * Creates a column schema for the database.
      *
-     * @param ColumnSchemaInterface $column The column schema information.
+     * This method may be overridden by child classes to create a DBMS-specific column schema.
+     *
+     * @param string $type The abstract data type.
+     * @param mixed ...$info The column information.
+     * @psalm-param array{unsigned?: bool} $info The set of parameters may be different for a specific DBMS.
+     *
+     * @return ColumnSchemaInterface
+     */
+    protected function createColumnSchema(string $type, mixed ...$info): ColumnSchemaInterface
+    {
+        $isUnsigned = !empty($info['unsigned']);
+        $phpType = $this->getColumnPhpType($type, $isUnsigned);
+
+        $column = $this->createColumnSchemaFromPhpType($phpType, $type);
+        $column->unsigned($isUnsigned);
+
+        return $column;
+    }
+
+    /**
+     * Get the PHP type from an abstract database type.
+     *
+     * @param string $type The abstract database type.
      *
      * @return string The PHP type name.
      */
-    protected function getColumnPhpType(ColumnSchemaInterface $column): string
+    protected function getColumnPhpType(string $type, bool $isUnsigned = false): string
     {
-        return match ($column->getType()) {
+        return match ($type) {
             // abstract type => php type
             SchemaInterface::TYPE_TINYINT => SchemaInterface::PHP_TYPE_INTEGER,
             SchemaInterface::TYPE_SMALLINT => SchemaInterface::PHP_TYPE_INTEGER,
-            SchemaInterface::TYPE_INTEGER => PHP_INT_SIZE === 4 && $column->isUnsigned()
+            SchemaInterface::TYPE_INTEGER => PHP_INT_SIZE !== 8 && $isUnsigned
                 ? SchemaInterface::PHP_TYPE_STRING
                 : SchemaInterface::PHP_TYPE_INTEGER,
-            SchemaInterface::TYPE_BIGINT => PHP_INT_SIZE === 8 && !$column->isUnsigned()
-                ? SchemaInterface::PHP_TYPE_INTEGER
-                : SchemaInterface::PHP_TYPE_STRING,
+            SchemaInterface::TYPE_BIGINT => PHP_INT_SIZE !== 8 || $isUnsigned
+                ? SchemaInterface::PHP_TYPE_STRING
+                : SchemaInterface::PHP_TYPE_INTEGER,
             SchemaInterface::TYPE_BOOLEAN => SchemaInterface::PHP_TYPE_BOOLEAN,
             SchemaInterface::TYPE_DECIMAL => SchemaInterface::PHP_TYPE_DOUBLE,
             SchemaInterface::TYPE_FLOAT => SchemaInterface::PHP_TYPE_DOUBLE,
@@ -390,6 +420,23 @@ abstract class AbstractSchema implements SchemaInterface
             SchemaInterface::TYPE_BINARY => SchemaInterface::PHP_TYPE_RESOURCE,
             SchemaInterface::TYPE_JSON => SchemaInterface::PHP_TYPE_ARRAY,
             default => SchemaInterface::PHP_TYPE_STRING,
+        };
+    }
+
+    protected function createColumnSchemaFromPhpType(string $phpType, string $type): ColumnSchemaInterface
+    {
+        return match ($phpType) {
+            SchemaInterface::PHP_TYPE_STRING => match ($type) {
+                SchemaInterface::TYPE_INTEGER => new BigIntColumnSchema($type, $phpType),
+                SchemaInterface::TYPE_BIGINT => new BigIntColumnSchema($type, $phpType),
+                default => new StringColumnSchema($type, $phpType),
+            },
+            SchemaInterface::PHP_TYPE_INTEGER => new IntegerColumnSchema($type, $phpType),
+            SchemaInterface::PHP_TYPE_DOUBLE => new DoubleColumnSchema($type, $phpType),
+            SchemaInterface::PHP_TYPE_BOOLEAN => new BooleanColumnSchema($type, $phpType),
+            SchemaInterface::PHP_TYPE_RESOURCE => new BinaryColumnSchema($type, $phpType),
+            SchemaInterface::PHP_TYPE_ARRAY => new JsonColumnSchema($type, $phpType),
+            default => new StringColumnSchema($type, $phpType),
         };
     }