+++ src/Schema/AbstractTableSchema.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Schema;
 
+use Yiisoft\Db\Schema\Column\ColumnSchemaInterface;
+
 use function array_keys;
 
 /**