+++ src/Schema/Column/ColumnSchemaInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Schema;
+namespace Yiisoft\Db\Schema\Column;
 
 /**
  * This interface defines a set of methods that must be implemented by a class that represents the column schema of a
@@ -167,9 +167,9 @@ interface ColumnSchemaInterface
     public function getExtra(): string|null;
 
     /**
-     * @return string The name of the column.
+     * @return string|null The name of the column.
      */
-    public function getName(): string;
+    public function getName(): string|null;
 
     /**
      * @return int|null The precision of the column.
@@ -243,6 +243,17 @@ interface ColumnSchemaInterface
      * @see unsigned()
      */
     public function isUnsigned(): bool;
+
+    /**
+     * Sets the name of the column.
+     *
+     * ```php
+     * $columns = [
+     *     'id' => $this->primaryKey()->name('id'),
+     * ];
+     * ```
+     */
+    public function name(string|null $name): void;
 
     /**
      * The PHP data type for representing the data stored in the column.