+++ src/Schema/TableSchemaInterface.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Schema;
 
+use Yiisoft\Db\Schema\Column\ColumnSchemaInterface;
+
 /**
  * Represents the metadata of a database table.
  *