+++ tests/Db/Schema/SchemaTest.php
@@ -65,33 +65,21 @@ final class SchemaTest extends AbstractSchemaTest
 
         $schema = $db->getSchema();
 
-        $columnBigInt = new ColumnSchema('bigint');
-        $columnBigInt->type('bigint');
-
-        $columnBoolean = new ColumnSchema('boolean');
-        $columnBoolean->type('boolean');
-
-        $columnInteger = new ColumnSchema('integer');
-        $columnInteger->type('integer');
-
-        $columnString = new ColumnSchema('string');
-        $columnString->type('string');
-
         $this->assertSame(
             'integer',
-            Assert::invokeMethod($schema, 'getColumnPhpType', [$columnBigInt]),
+            Assert::invokeMethod($schema, 'getColumnPhpType', ['bigint']),
         );
         $this->assertSame(
             'boolean',
-            Assert::invokeMethod($schema, 'getColumnPhpType', [$columnBoolean]),
+            Assert::invokeMethod($schema, 'getColumnPhpType', ['boolean']),
         );
         $this->assertSame(
             'integer',
-            Assert::invokeMethod($schema, 'getColumnPhpType', [$columnInteger]),
+            Assert::invokeMethod($schema, 'getColumnPhpType', ['integer']),
         );
         $this->assertSame(
             'string',
-            Assert::invokeMethod($schema, 'getColumnPhpType', [$columnString]),
+            Assert::invokeMethod($schema, 'getColumnPhpType', ['string']),
         );
     }